requirements.txt
Django==2.2
requests
djangorestframework==3.11.0